# Removing optimization that never got used