--- a/include/ada/unicode.h
+++ b/include/ada/unicode.h
@@ -94,7 +94,7 @@ ada_really_inline constexpr bool contains_forbidden_domain_code_point(
  * then the second bit is set to 1.
  * @see https://url.spec.whatwg.org/#forbidden-domain-code-point
  */
-ada_really_inline constexpr bool contains_forbidden_domain_code_point_or_upper(
+ada_really_inline constexpr uint8_t contains_forbidden_domain_code_point_or_upper(
     const char* input, size_t length) noexcept;
 
 /**