We had an accidental casting of a byte value to a bool which was disabling an optimization in our code. It turns out that this 'fast' path was never used.

Removing it does not measurably affect the performance and it removes lines of code we can do without.

This routine could be revisited later for more optimizations.

Current main branch on benchdata (Apple M1, LLVM 14):

```
BasicBench_AdaURL_aggregator_href   16168863 ns     16168814 ns           43 GHz=3.4071 cycle/byte=6.30681 cycles/url=547.804 instructions/byte=28.5425 instructions/cycle=4.52566 instructions/ns=15.4194 instructions/url=2.47918k ns/url=160.783 speed=537.336M/s time/byte=1.86103ns time/url=161.648ns url/s=6.18629M/s
```

With this PR

```
BasicBench_AdaURL_aggregator_href   15900795 ns     15900750 ns           44 GHz=3.49309 cycle/byte=6.3603 cycles/url=552.451 instructions/byte=28.1538 instructions/cycle=4.42649 instructions/ns=15.4621 instructions/url=2.44542k ns/url=158.155 speed=546.395M/s time/byte=1.83018ns time/url=158.968ns url/s=6.29058M/s
```