@lemire 2023-05-18
@anonrig This PR will not require a new release. There is no bug and this PR does not improve the performance. It is just basic hygiene.